FROMLIST: android: binder: Move buffer out of area shared with user space
(from https://patchwork.kernel.org/patch/9928607/)

Binder driver allocates buffer meta data in a region that is mapped
in user space. These meta data contain pointers in the kernel.

This patch allocates buffer meta data on the kernel heap that is
not mapped in user space, and uses a pointer to refer to the data mapped.

Also move alloc->buffers initialization from mmap to init since it's
now used even when mmap failed or was not called.

Bug: 36007193
Change-Id: Iff0448a8a7fa0171a62bf0a18fb05bbf29d34020
Signed-off-by: Sherry Yang <sherryy@android.com>