prima: Enable Wifi channel bonding mode unconditionally
Wifi channel bonding is known to improve the internet speed. But
it depends on userspace configuration, so we need to tweak it with
something like Magisk.

Avoid this by unconditionally set wifi channel bonding mode to
true in kerenl. Note that it is a _HACK_, so only use this when
you need it.

Signed-off-by: Wang Han <416810799@qq.com>