usb: gadget: FunctionFS: Remove VLAIS usage from gadget code
The use of variable length arrays in structs (VLAIS) in the Linux Kernel
code
precludes the use of compilers which don't implement VLAIS (for instance
the
Clang compiler). This alternate patch calculates offsets into the
kmalloc-ed
memory buffer using macros. The previous patch required multiple kmalloc
and
kfree calls. This version uses "group" vs "struct" since it really is
not a
struct and is essentially a group of VLA in a common allocated block.
This
version also fixes the issues pointed out by Andrzej Pietrasiewicz and
Michal Nazarewicz.

Signed-off-by: Mark Charlebois <charlebm@gmail.com>
Signed-off-by: Behan Webster <behanw@converseincode.com>

[elimination of miexed declaration and code, checkpatch cleanup]
[fixes after Michal's review]
Signed-off-by: Andrzej Pietrasiewicz <andrzej.p@samsung.com>
Signed-off-by: Kyungmin Park <kyungmin.park@samsung.com>
Acked-by: Michal Nazarewicz <mina86@mina86.com>
Signed-off-by: Felipe Balbi <balbi@ti.com>
[md: Backport to 3.10]
Signed-off-by: mortezadadgar <mortezadadgar97@gmail.com>