DT:Remove CPU oc and restore stable values
we are getting a lot of EDAC failures in our logs on some kenzo CPUs and
its because of low voltage and low voltage on cpu causes faults to occur
in cpu registers. even though its correctable error it causes spikes in
heavy load tasks and its not pretty to see that and we are'nt saving a
huge deal of power so lets reduce the voltage a bit.